fix(security): SMI-5922 — resolve CodeQL js/clear-text-logging false positive
GitHub Advanced Security flagged 2 high-severity js/clear-text-logging
alerts in scripts/e2e-registry-roundtrip.ts's record() sink, traced to
noAuthFetch.status and noAuthFetch.cacheControl flowing into a log
line. Verified false positive: CodeQL's source heuristic treats ANY
property read off a fetch response from an auth-adjacent endpoint as
"sensitive," but the actual values here are a plain HTTP status code
and a Cache-Control header string -- never a credential. Every
`detail` value logged anywhere in this file is a status code, header
value, or denial-response error body (rawPrivateRegistryGet() is only
ever called on the 401/403/404 denial paths, never with real skill
content), so no secret or PII can reach these log lines. Suppressed
with inline `codeql[js/clear-text-logging]` comments plus a
justification comment, at both flagged sink lines.

Also splits scripts/e2e-registry-roundtrip.ts into a fourth file
(.helpers.ts, holding signIn/rawPrivateRegistryGet/runCliInstall) --
the security-fix comments pushed the main file to 510 lines, over the
500-line gate; extracting the three standalone HTTP/CLI-subprocess
helpers (matching the seed script's own .helpers.ts split) brings it
to 384 with real headroom. Added the new file to the workflow's
prod-ref grep gate and PR trigger paths.

Re-verified: FULL_TYPECHECK=1 typecheck clean, lint clean,
audit:standards 78/0/8(pre-existing), full test suite 955 files /
15,555 tests / 0 failures, prod-ref grep gate clean, workflow YAML
parses.

Part of SMI-5922 (docs/internal/implementation/smi-5922-private-registry-e2e-validation.md).
